local-ollama-version/requirements.txt
streamlit==1.52.2
crewai==1.7.2
linkup-sdk==0.9.0
python-dotenv==1.1.1
pydantic==2.11.10
fpdf2==2.7.9
